Decimal

+++ BelezaAPI/BelezaAPI/Dominio/Servico.cs
@@ -4,7 +4,7 @@ public class Servico
 {
 	public string Nome { get; set; }
 	public int Id { get; set; }
-	public int Preco { get; set; }
+	public decimal Preco { get; set; }
 	public bool Disponibilidade { get; set; }
     public int DuracaoMin { get; set; }
 }
